Removed separate login view and added the textboxes to main login/home view
