+++ app/src/main/res/layout/activity_main.xml
@@ -12,8 +12,24 @@
     <ImageView
         android:id="@+id/imageView3"
         android:layout_width="wrap_content"
-        android:layout_height="412dp"
+        android:layout_height="300dp"
         app:srcCompat="@drawable/d98dde45d242734" />
+
+    <EditText
+        android:id="@+id/username"
+        android:layout_width="300dp"
+        android:layout_height="wrap_content"
+        android:textSize="25sp"
+        android:fontFamily="serif-monospace"
+        android:hint="Username"/>
+
+    <EditText
+        android:id="@+id/password"
+        android:layout_width="300dp"
+        android:layout_height="wrap_content"
+        android:textSize="25sp"
+        android:fontFamily="serif-monospace"
+        android:hint="Password"/>
 
     <Button
         android:id="@+id/login"
