+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,3 @@
-#Wed Sep 01 21:00:11 PDT 2021
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
